修改应用信息

+++ services/toucan-shopping-admin-auth/src/main/java/com/toucan/shopping/admin/auth/mapper/AppMapper.java
@@ -18,5 +18,5 @@ public interface AppMapper {
 
     List<App> queryListPage(AppPageInfo appPageInfo);
 
-    int deleteById(Integer id);
+    int deleteById(Long id);
 }
